InvoiceQr 컴포넌트를 추출한다
